sql/catalog/descs: extract a file and concept for kvDescriptors
This allows us to isolate, a bit, the concept of reading from the
store and caching that read data.

Release note: None
